# docs/swagger.yaml
components:
  schemas:
...
    github_com_leseb_openresponses-gw_pkg_core_schema.DeleteFileResponse:
      properties:
        deleted:
          description: Always true
          type: boolean
        id:
          description: File ID
          type: string
        object:
          description: Always "file.deleted"
          type: string
      type: object
...
    github_com_leseb_openresponses-gw_pkg_core_schema.File:
      properties:
        bytes:
          description: File size in bytes
          type: integer
        created_at:
          description: Unix timestamp
          type: integer
        filename:
          description: Original filename
          type: string
        id:
          description: 'Format: "file_{uuid}"'
          type: string
        mime_type:
          description: MIME type
          type: string
        object:
          description: Always "file"
          type: string
        purpose:
          description: 'Purpose: "assistants", "vision", "batch", "fine-tune"'
          type: string
        status:
          description: 'Status: "uploaded", "processed", "error"'
          type: string
      type: object
...
    github_com_leseb_openresponses-gw_pkg_core_schema.ListFilesResponse:
      properties:
        data:
          description: Array of files
          items:
            $ref: '#/components/schemas/github_com_leseb_openresponses-gw_pkg_core_schema.File'
          type: array
          uniqueItems: false
        first_id:
          description: ID of first item
          type: string
        has_more:
          description: Whether there are more results
          type: boolean
        last_id:
          description: ID of last item
          type: string
        object:
          description: Always "list"
          type: string
      type: object
...
info:
  contact:
    name: Open Responses Gateway
    url: https://github.com/leseb/openresponses-gw
  description: |-
    100% Open Responses Specification Compliant Gateway.
    Based on: https://github.com/openresponses/openresponses

    This gateway provides:
    - **Core API**: Full Open Responses spec compliance (POST /v1/responses)
    - **Extended APIs**: Conversations, Prompts, Files, Vector Stores, Connectors
    - **Dual Mode**: Standalone HTTP server or Envoy ExtProc integration

    Streaming: All 24 event types from Open Responses spec
    Request Echo: All request parameters returned in response
    Multimodal: Support for text, images, files, video
  title: Open Responses Gateway API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /v1/files:
    get:
      parameters:
      - description: Cursor for pagination
        in: query
        name: after
        schema:
          type: string
      - description: Cursor for pagination (backwards)
        in: query
        name: before
        schema:
          type: string
      - description: Number of items (1-100, default 50)
        in: query
        name: limit
        schema:
          type: integer
      - description: 'Sort order: asc or desc (default desc)'
        in: query
        name: order
        schema:
          type: string
      - description: Filter by purpose
        in: query
        name: purpose
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_leseb_openresponses-gw_pkg_core_schema.ListFilesResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Internal Server Error
      summary: List files
      tags:
      - Files
    post:
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              oneOf:
              - title: file
                type: file
              - title: purpose
                type: string
          multipart/form-data:
            schema:
              type: object
        description: 'File to upload | Purpose: assistants, vision, batch, or fine-tune'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_leseb_openresponses-gw_pkg_core_schema.File'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Internal Server Error
      summary: Upload file
      tags:
      - Files
  /v1/files/{id}:
    delete:
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_leseb_openresponses-gw_pkg_core_schema.DeleteFileResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Not Found
      summary: Delete file
      tags:
      - Files
    get:
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_leseb_openresponses-gw_pkg_core_schema.File'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Not Found
      summary: Get file
      tags:
      - Files
  /v1/files/{id}/content:
    get:
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/octet-stream:
              schema:
                type: file
          description: OK
        "400":
          content:
            application/octet-stream:
              schema:
                additionalProperties: {}
                type: object
          description: Bad Request
        "404":
          content:
            application/octet-stream:
              schema:
                additionalProperties: {}
                type: object
          description: Not Found
        "500":
          content:
            application/octet-stream:
              schema:
                additionalProperties: {}
                type: object
          description: Internal Server Error
      summary: Get file content
      tags:
      - Files
...
servers:
- description: Local development server
  url: http://localhost:8080
...